Restructure basis function modules (#63)
* reorganized modules

* fixed formatting and static typing issues

* updated test imports

* relative imports to absolute imports
