+++ tests/gpflux/layers/basis_functions/fourier_features/test_quadrature.py
@@ -23,8 +23,8 @@ import gpflow
 from gpflow.quadrature.gauss_hermite import NDiagGHQuadrature
 from gpflow.utilities.ops import difference_matrix
 
-from gpflux.layers.basis_functions.fourier_features import QuadratureFourierFeatures
-from gpflux.layers.basis_functions.fourier_features.utils import QFF_SUPPORTED_KERNELS
+from gpflux.layers.basis_functions.fourier_features.quadrature import QuadratureFourierFeatures
+from gpflux.layers.basis_functions.fourier_features.quadrature.gaussian import QFF_SUPPORTED_KERNELS
 
 
 @pytest.fixture(name="n_dims", params=[1, 2, 3])
