+++ tests/gpflux/layers/basis_functions/fourier_features/test_random.py
@@ -21,12 +21,12 @@ from tensorflow.python.keras.utils.kernelized_utils import inner_product
 
 import gpflow
 
-from gpflux.layers.basis_functions.fourier_features import (
+from gpflux.layers.basis_functions.fourier_features.random import (
     OrthogonalRandomFeatures,
     RandomFourierFeatures,
     RandomFourierFeaturesCosine,
 )
-from gpflux.layers.basis_functions.fourier_features.utils import RFF_SUPPORTED_KERNELS
+from gpflux.layers.basis_functions.fourier_features.random.base import RFF_SUPPORTED_KERNELS
 
 
 @pytest.fixture(name="n_dims", params=[1, 2, 3, 5, 10, 20])
